==> exp/04_TFIDF_and_countvect_with_POS.ipynb <==
In this and following notebooks (05.Scispacy_for_NER_and_similarities) we tried tf-idf, count-vectorizer and SciBERT models to identify named entities or concepts in Mathematics for Machine Learning Textbook. 

import pandas as pd
import re
import nltk
import numpy as np
#nltk.download('all')
#nltk.download('punkt')
#nltk.download('wordnet')
#nltk.download("omw-1.4")
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer

with open ("../dat/parsed_books/mml-book.txt", encoding = "utf8") as f:
    text = f.readlines()
text = "".join(text)
print(len(text))
text = text.replace('©2021 M. P. Deisenroth, A. A. Faisal, C. S. Ong. Published by Cambridge University Press (2020).'," ")
text = text.replace("\n©by M. P. Deisenroth, A. A. Faisal, and C. S. Ong, 2021. https://mml-book.com ."," ")
text = text.replace('Draft (2022-01-11) of “Mathematics for Machine Learning”. Feedback: https://mml-book.com .'," ")
text = text.replace("This material is published by Cambridge University Press as Mathematics for Machine Learning by\nMarc Peter Deisenroth, A. Aldo Faisal, and Cheng Soon Ong (2020). This version is free to view\nand download for personal use only. Not for re-distribution, re-sale, or use in derivative works."," ")
print(len(text))


part1 = text.split("\nPart II\nCentral Machine Learning Problems")[0]
len(part1)

chapters = ["\n1\nIntroduction and Motivation\n","\n2\nLinear Algebra\n","\n3\nAnalytic Geometry\n","\n4\nMatrix Decompositions\n",
            "\n5\nVector Calculus\n","\n6\nProbability and Distributions\n","\n7\nContinuous Optimization\n"]

chapters_dict ={}
rest_chapters = part1 ## .split(chapters[0])[1]

# Chapter 0 (not in index): intro
# chapter 1: ...
# ...
# Chapter n : ....
for idx,chapter_name in enumerate(chapters):
    text_splitted = rest_chapters.split(chapter_name)
    if idx ==0:
        pass
    else:
        chapters_dict[chapters[idx-1][3:-1]] = text_splitted[0]
    print("chapter: "+ chapter_name)
    rest_chapters = text_splitted[1]
chapters_dict[chapters[idx][3:-1]] = text_splitted[1]

data = pd.DataFrame.from_dict(chapters_dict, orient='index',columns=["text"])
data

def prep(text,delete_stopwords=True):
    #text = re.sub("\n"," ",text)
    #text = re.sub("\d+", "", text)

    # words only
    regex=u"[A-Za-z]+"
    regex = re.compile(regex)
    text = " ".join(regex.findall(text))
    # lower
    text = text.lower()
    lemmatizer = WordNetLemmatizer()
    word_tokens = nltk.word_tokenize(text)
    # remove stop words + lemmatization
    if delete_stopwords:
        stop_words = set(stopwords.words('english'))
        text = " ".join([lemmatizer.lemmatize(w) for w in word_tokens if not w in stop_words])
    else:
        text = " ".join([lemmatizer.lemmatize(w) for w in word_tokens])  
    return text

data["text_prep"] = data.text.apply(prep)

#### TF-IDF TRANSFORMATION

# !!! tf-idf 

tfidf = TfidfVectorizer(ngram_range=(1,3))
tfidf_matrix = tfidf.fit_transform(data.text_prep.values)
print(tfidf_matrix.shape)

vocab = {v:k for k,v in tfidf.vocabulary_.items()}

data["tfidf_topwords"]=None
for n,chapter in enumerate(data.index):
    s = ""
    for idx in np.argsort(tfidf_matrix.A[n,:])[::-1][:50]:
        s += ", "+vocab[idx]
    data.loc[chapter,"tfidf_topwords"] = s[2:]

    print(chapter+": \n ", data.loc[chapter,"tfidf_topwords"])

#### SIMPLE BAG OF WORDS (COUNTVECTORIZER) + POS

from collections import Counter
Counter(data.text_prep['Linear Algebra'].split(" ")).most_common()[:15]

# POS tagging 
data["text_prep_stopwords_incl"] = data.text.apply(prep,delete_stopwords=False)

def filter_by_POS(text):
    tokens = nltk.word_tokenize(text)
    pos = nltk.pos_tag(tokens)
    text = [p[0] for p in pos if p[1] in ["VBG","RB","RBR","RBS","JJ","JJR","JJS","NN","NNS"]]
    return " ".join(text)

data["text_nouns_adj_adv"] = data["text_prep_stopwords_incl"].apply(filter_by_POS)

data["text_nouns_adj_adv"]["Introduction and Motivation"]

countvect = CountVectorizer(analyzer='word',ngram_range=(1, 2))
countvect_matrix = countvect.fit_transform(data.text_nouns_adj_adv.values)
print(countvect_matrix.shape)

vocab = {v:k for k,v in countvect.vocabulary_.items()}

data["countvect_topwords"]=None
for n,chapter in enumerate(data.index):
    s = ""
    for idx in np.argsort(countvect_matrix.A[n,:])[::-1][:50]:
        s += ", "+vocab[idx]
    data.loc[chapter,"countvect_topwords"] = s[2:]

    print(chapter+": \n ", data.loc[chapter,"countvect_topwords"])
