==> exp/05_Scispacy_for_NER_and_similarities.ipynb <==
In this and previous notebooks (04_TFIDF_and...) we tried tf-idf, count-vectorizer and SciBERT models to identify named entities or concepts in Mathematics for Machine Learning Textbook. 

## 1. ScispaCy

# 1. Create an environment for scispacy
#conda create -n scispacy python=3.9

# 2. Load a model trained on scientific publications
# #! pip install https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.1/en_core_sci_sm-0.5.1.tar.gz
# ! pip install https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.1/en_core_sci_scibert-0.5.1.tar.gz 
! pip install https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.1/en_core_sci_md-0.5.1.tar.gz
# NOTE: scibert doesn't have vectors

# 3. Install spacy
#! pip install spacy


#! pip install pandas
import pandas as pd
import spacy

data = pd.read_json("../dat/parsed_books/mml.json")
data

nlp = spacy.load("en_core_sci_md")
doc=nlp(data.text["Linear Algebra"])
entities_la = doc.ents

df_entities = pd.DataFrame(columns = ["entity","lemma","start_char","end_char","position in text","embedding"])
for n, entity in enumerate(entities_la):
    df_entities.loc[n,:] = [entity, entity.lemma_.lower(), entity.start_char, entity.end_char, entity.start, entity.vector]
df_entities.head(5)

df_entities.loc[df_entities.lemma == "vector", "embedding"] # shape of embedding is 200

from collections import defaultdict
freqs = defaultdict(int)
for w in df_entities.lemma:
    freqs[w] += 1

freqs = dict(sorted(freqs.items(), key=lambda item: item[1], reverse = True))
freqs

# calculating similarity
import numpy as np
from numpy.linalg import norm
def find_most_similar(word, df_entities):
    try:
        word_embedding = df_entities.loc[df_entities.lemma == word,"embedding"].mean()
        rest = (df_entities.loc[df_entities.lemma != word,["lemma","embedding"]]
                .groupby("lemma").mean()
                .reset_index()
            )
        rest_embeddings = rest["embedding"].values
        rest_entities = rest["lemma"].values

        cosine_similarity = []
        for embedding in rest_embeddings:
            cosine_similarity.append(embedding @ word_embedding / (norm(embedding)*norm(word_embedding)+10e-10))
        best = np.array(cosine_similarity).argsort()[-5:][::-1]
        return rest_entities[best]
    except:
        print("This word is not included in the dataset")
    

find_most_similar("matrix", df_entities)

find_most_similar("vector", df_entities)

find_most_similar("determinant", df_entities)

find_most_similar("inverse", df_entities)

#### Conclusion: 

1. NER: en_core_sci_md model identifies entities quite well, even if it was trained on biomedical data. However, some subsequent filtering is required.

2. Cosine similarity between word embeddings does not allow finding most similar concepts supposedly for the following reasons:

- corpus is small

- noise in data like '.finite-dimensional vector\nspacesvandware' and etc.

- need to check "the meaning" of vectors, maybe I misused them

- averages of the vectors may not represent a "real" vector/embedding of a word in the vector space.

## SciBert, only transformers library

from transformers import BertTokenizer, BertForTokenClassification, pipeline

tokenizer = BertTokenizer.from_pretrained('allenai/scibert_scivocab_uncased')
model = BertForTokenClassification.from_pretrained('allenai/scibert_scivocab_uncased')
nlp = pipeline('ner', model=model, tokenizer=tokenizer)

doc = nlp(data.text["Linear Algebra"][:1000])

doc[:5]
