==> exp/07_Parallel_Wikification.ipynb <==
# Wikification
This notebooks was created to find a fast way of Wikipedia article annotation.

import multiprocessing as mp
print("Number of processors: ", mp.cpu_count())

from PyPDF2 import PdfReader
import os
import pandas as pd
import numpy as np

import sys
sys.path.append('../src/')
import wikifier

path = "../dat/books/all_books/"
data = pd.DataFrame(columns = ["file","text"])

book_files = os.listdir(path)

for j, book in enumerate(book_files):
    reader = PdfReader(path + book)
    print("Extracting " + book)
    text = ""
    k = 0
    for page in reader.pages:
        k+=1
        if k%50==0: print("Extract page" + str(k))
        text += str(page.extract_text()) + " "

    data = pd.concat([data, pd.DataFrame.from_records([{'file': book, 'text': text}])])
data.reset_index()

data.text.apply(len)

# exclude Math14, Strang

data.text[8]

batch_size = 5000
num_threads = mp.cpu_count()

all_annotations = []
for i in range(2,3):
    text = data.iloc[i].text

    pool = mp.Pool(num_threads)
    print("Length: " + str(len(text)))
    annotations = [pool.apply(wikifier.get_annotations, args=[text[i*batch_size:(i+1)*batch_size], i*batch_size])
        for i in range(int(np.floor(len(text)/batch_size)))]
    pool.close()
    all_annotations = sum(annotations, [])

annos = [{
    "title": a["title"],
    "supports": [{'chFrom': s['chFrom'], 'chTo': s['chTo']} for s in a["support"] if s['pageRank']>0.001]
    } for a in all_annotations]
[a for a in annos if len(a['supports'])]

len(all_annotations)
import pickle
file = open("../dat/annotations_Denton.pkl", 'wb')
pickle.dump(all_annotations, file)
file.close()

import pickle
file = open("../dat/annotations/Beezer_First_Course.pkl", 'rb')
annos = pickle.load(file)

len(annos)
