==> exp/11_PyMuPDF_data_prep.ipynb <==
In this notebook, we preprocessed texts that were parsed from textbooks by using PyMuPDF library for further analysis. This includes removing unnecessary symbols and page tokenization.

import fitz
import pandas as pd
import os
import unicodedata
from IPython.display import clear_output
import time

short_words=["a",'i',"of","on","in","ok","so","be","to","me",'we','he','ai','am','an','as','at','bi','by','ed','en','hi','id','if','in','is','it','my','or','os','pi','us','no']
short_words = short_words + [word.upper() for word in short_words] + [word.capitalize() for word in short_words]
def has_numbers(token):
    return any(char.isdigit() for char in token)

data = pd.DataFrame(columns=["pages"])

folder = "../dat/books/selected10/"


start_pages = []
end_pages = []
for book in os.listdir(folder):
    doc = fitz.open("../dat/books/selected10/"+book)

    # table of content
    toc = doc.get_toc()
    tocs_1level = [item for item in toc if item[0]==1]
    print(tocs_1level)
    time.sleep(4)

    # input the number of page where the main content starts
    start = int(input())
    start_pages.append(start)

    # input the number of the next page after the main content
    end = int(input())
    end_pages.append(end)

    # read the text from pdf and remove unnecessary content and noise
    text=[]
    for i,page in enumerate(doc):
        if (i+1>=start) and (i+1<end):

            page_text = page.get_text()

            page_text = unicodedata.normalize("NFKD", page_text)
            page_text = page_text.encode("ascii", "ignore")
            page_text = page_text.decode()
            page_text = page_text.replace("-\n","")
            page_text = page_text.replace("\n"," ")

            text.append(page_text)
        else:
            text.append("")
    

    cleaned_text=[]
    for page in text:
        cleaned_tokens = []
        for token in page.split(" "):
            if not ((has_numbers(token)) or ((len(token) <=2) and (token not in short_words))):
                cleaned_tokens.append(token)
        cleaned_text.append(" ".join(cleaned_tokens))
    
    data.loc[book[:-4],"pages"]=cleaned_text
    clear_output(wait=True)

data

data.loc['Beezer_First_Course','pages'][100]

start_pages

end_pages

data.to_json("../dat/parsed_books/parsed_books.json")
