==> exp/14_ChatGPT_dependencies.ipynb <==
# Query ChatGPT
In this notebook, requests should be send to ChatGPT to query it for concept dependencies.

import json
with open('../config.json', 'r') as f:
    config = json.load(f)

from revChatGPT.revChatGPT import Chatbot

#config = {
#    "email": "",
#    "password": "",
#    #"session_token": "<SESSION_TOKEN>", # Deprecated. Use only if you encounter captcha with email/password
#    #"proxy": "<HTTP/HTTPS_PROXY>"
#}

chatbot = Chatbot(config, conversation_id=None)
chatbot.refresh_session() # You need to log in on the first run

answer = chatbot.get_chat_response("Do I have to understand Linear Transformation Addition to understand Injective Linear Transformation?", output="text")
print(answer)

answer = chatbot.get_chat_response("In linear algebra, do I have to understand Equivalent Systems to understand joint variation?", output="text")
print(answer['message'])

import pandas as pd
import time

index = pd.read_csv("../dat/index/CollegeAlgCoreq-WEB.csv")

print(index.iloc[1:3,1])

subset = index.sample(5).reset_index()
subset=subset.drop(subset.columns[0:2], axis=1)

subset

gpt_resp = []

for i in range(5):
    target_concept = subset.iloc[i,0]
    for j in range(5):
        concept_dep = subset.iloc[j,0]
        # ask Gpt Chat
        question = "In linear algebra, do I have to understand " + concept_dep + "to understand " + target_concept + "?"
        answer = chatbot.get_chat_response(question, output="text")
        print(answer['message'])
        gpt_resp.append({
            "target_concept": target_concept,
            "concept_dependency": concept_dep,
            "question": question,
            "answer": answer['message']
        })
        time.sleep(20)

gpt_resp

df_gpt = pd.DataFrame(gpt_resp)
print(df_gpt.columns)
df_gpt['target_concept'] = df_gpt['target_concept'].str.strip()
df_gpt['concept_dependency'] = df_gpt['concept_dependency'].str.strip()
df_gpt.to_csv("../dat/chatgpt/sample_resps.csv")
