==> exp/16a_Index_to_Wiki.ipynb <==
# Index entry disambigation
Here, the texbook index entries are mapped to Wikipedia articles using the Wikisearch disambiguation method.

import os
import pandas as pd
from tqdm import tqdm
from collections import Counter
import sys
sys.path.append('../src/')
import wiki_api
from importlib import reload
wiki_api = reload(wiki_api)

path = "../dat/index/"

def unique_sorted(pages):
    pages = list(set(pages.split(","))) # remove duplicates
    pages = [p.strip() for p in pages]
    while "nan" in pages: pages.remove("nan")
    pages.sort(key=int)
    return ", ".join(pages)

disamb = wiki_api.disambiguate("injection", None, "Mathematics", True)


for book in os.listdir(path)[:6]:
    index = pd.read_csv(path + book)
    for title in index.concept:
        print(f"Index entry: {title}")
        disamb = wiki_api.disambiguate(str(title), None, "Mathematics", True)
        print(f"Wiki result: {disamb}")


for book in os.listdir(path):
    index = pd.read_csv(path + book)
    index.wiki_concept = [wiki_api.disambiguate(str(title), None, "Mathematics") for title in tqdm(index.concept)]
    
    index.pages = index.pages.apply(str) # to prepare join
    wiki_index = index.groupby("wiki_concept").agg({'pages': ', '.join})
    wiki_index.pages = wiki_index.pages.apply(unique_sorted)
    wiki_index.to_csv("../dat/index_by_wiki/"+ book)

## Descriptive stats

path = "../dat/index_by_wiki/"
wiki_concepts = []
for book in os.listdir(path):
    index = pd.read_csv(path + book)
    wiki_concepts.extend(index.wiki_concept)

count = Counter(wiki_concepts)

Number of unique concepts

len(count)

Concepts that appear in at least 4 books

common_concepts = [c for c in count if count[c]>=5]
len(common_concepts)

print(common_concepts)
