==> exp/21_Query_LLM.ipynb <==
# LLM Concept Dependency Graph
In this notebook, T0pp is prompted in the fashion of our output re-feeding method to create the dependency graph.

import numpy as np
import pandas as pd
import requests
import json
import ast

import sys
sys.path.append('../src/')
import wiki_api
import data_layer
import graph_api


host = "localhost"
port = "8100"

# T0PP

## Example outputs

get_answer(f"Simple definition of a hyperplane:")

get_answer(f"What mathematical concepts are mentioned here? A hyperplane is a plane in three-dimensional space that is perpendicular to all other planes and tangent to all other lines.")

get_answer(f"In mathematics, what is matrix?")

get_answer(f"Give me a detailled definition of Kernel (algebra)")

get_answer(f"In linear algebra, what is matrix transpose?")

get_answer(f"Germany // Berlin\nFrance // Paris\nRussia // Moscow\nItaly - Rome\n USA //")

## Dep Extraction

def get_answer(prompt):
    url = f"http://{host}:{port}?prompt={prompt}"
    response = requests.request("GET", url)
    result = json.loads(response.text)
    return result['response'].replace("<pad> ","").replace("</s>", "")

def get_deps(i, concept):
    if i % 5 == 0: print(i)
    definition = get_answer(f"What is the mathematical definition of {concept}?")
    response = get_answer(f"What mathematical concepts are mentioned here: {definition}")
    deps = list(map(lambda x: x.strip(), response.split(",")))
    return deps

_, wiki_concepts = data_layer.read_index_and_wiki_concepts()
wiki_concepts = list(wiki_concepts)

#concepts = np.random.choice(wiki_concepts, 100)
deps = [get_deps(i, concept) for i, concept in enumerate(wiki_concepts)]
df_concepts = pd.DataFrame({"concept": wiki_concepts, "deps": deps})
df_concepts.to_csv("../dat/llm/t0pp_output.csv",index=False)

#df_concepts = pd.read_csv("../dat/llm/t0pp_output.csv")
#df_concepts.deps = df_concepts.deps.apply(ast.literal_eval)
#df_concepts.to_json("../dat/llm/t0pp_output.json")

link_set = set(wiki_concepts)
df_concepts['dep_articles'] = [print(i) or [wiki_api.disambiguate(dep, link_set, "Mathematics") for dep in concept_deps] for i, concept_deps in enumerate(df_concepts.deps)]

#df_concepts = pd.read_csv("../dat/llm/df_concepts.csv")
#df_concepts.deps = df_concepts.deps.apply(ast.literal_eval)
#df_concepts.dep_articles = df_concepts.dep_articles.apply(ast.literal_eval)
#df_concepts.to_json("../dat/llm/t0pp_graph.json")

def delete_reflexive(concept, dep_articles):
    return list(filter(lambda x: x!=concept, dep_articles))
df_concepts.dep_articles = [delete_reflexive(concept, dep_articles) for concept, dep_articles in zip(df_concepts.concept, df_concepts.dep_articles)]

df_concepts.to_json("../dat/llm/graph.json")
df_concepts.to_csv("../dat/llm/df_concepts.csv",index=False)

#print(list(zip(df_concepts.concept, df_concepts.deps, df_concepts.dep_articles)))

# Galactica

def get_answer(prompt):
    url = f"http://{host}:{port}?prompt={prompt}"
    response = requests.request("GET", url)
    result = json.loads(response.text)
    return result['response'] #.replace("<pad> ","").replace("</s>", "")

get_answer(f"This is a list of mathematical concepts that are required to define matrix:\n")

get_answer(f"Brief definition of matrix:\n")

get_answer(f"A matrix is a rectangular array of numbers, symbols, or expressions arranged in rows and columns.\n\nQ:What are the mathematical objects used in the above definition?\n\nA:")

get_answer(f"Brief definition of hyperplane:\n")

get_answer(f"A hyperplane is a set of points in a vector space that are equidistant from a given point.\n\nQ: What three objects are mentioned in the abstract above?\n\nA:'")

get_answer(f"This is a list of concepts in the definition of Convex set:\n\n")

indices, wiki_concepts = data_layer.read_index_and_wiki_concepts()

for concept in np.random.choice(wiki_concepts, 20):
    print("CONCEPT: " + concept)
    context = wiki_api.get_first_paragraph(concept)
    response = get_answer(f"{context}\n\nQ: What mathematical concepts are mentioned in the abstract above?\n\nA:")
    print(response)
