==> exp/23_validation_dataset.ipynb <==
# Creation of validation dataset
Here, the subsample for the manual validation is created.

import pandas as pd
from random import choice
import sys
sys.path.append('../src/')
import data_layer

MIN_OCCURENCES=5 # minimum number of occurences in textbooks

df_wiki = pd.read_csv("../dat/wiki/graph.csv")
df_llm = pd.read_json("../dat/llm/graph.json")
df_txb = pd.read_json("../dat/textbooks/graph.json").iloc[:,[0,3]]

_, _, df_count = data_layer.read_index_and_wiki_concepts(include_counts=True)
df_count.sort_values('count', ascending=False)


df_count[df_count['count']>=MIN_OCCURENCES]

n=100
rnd_st=123
concepts = []
deps = []

for df in [df_llm, df_wiki,df_txb]:
    df = pd.merge(df, df_count, on="concept")
    df = df.loc[(df.deps.str.len() > 0) & (df['count']>=MIN_OCCURENCES)]
    samp = df.sample(n,random_state=rnd_st)
    concepts.extend(samp['concept'])
    deps.extend([choice(i) for i in samp['dep_articles']])
    
len(deps) == len(concepts)

samples = pd.DataFrame({"concept":concepts,"dependency":deps,"source":["llm"]*n+['wiki']*n+['textbook']*n})
samples = samples.sample(frac=1).reset_index(drop=True)
samples['source'].to_csv('../dat/validation/source_column.csv')
samples.loc[:,['concept','dependency']].to_excel('../dat/validation/deps_to_validate.xlsx')
